+++ NEWS.md
@@ -4,10 +4,11 @@
 * Added palette based on Red Dead Redemption II ("cowboy")
 * Added palette based on Persona 4 Golden ("p4g")
 * Added palette based on Legend of Zelda: Ocarina of Time ("ocarina")
+* Now has a function to generate continuous palettes. 
 
 # gameR 0.0.1
 
 * First release of the package
 * Added a `NEWS.md` file to track changes to the package.
-* Add discrete palettes based on Cuphead, Cyberpunk 2077, Fallout, Gris, Rayman,
-  Spiritfarer, Splatoon, Super Mario Bros, and World of Warcraft.
+* Added discrete palettes based on Cuphead, Cyberpunk 2077, Fallout, Gris,
+  Rayman, Spiritfarer, Splatoon, Super Mario Bros, and World of Warcraft.